Nasty turd of a memory-corruption.  OF stub did not relocate ofh_cchan in ohf_cons_init(), this resulted in a wild write to the text segment of a Linux kernel loaded at 0.  (Of course finding this was very, very quick easy and painless.)
Fix the ohf_cons_init name.
Force relocation of ofh_cchan in ofh_cons_init.

Break ofh_cons_read, ofh_cons_write --- they have no way of knowing of
how to relocate ofh_cchan, because the base addrss is not passed to
them.  In fact, it can't be passed to the because they are called
through leap(), and leap() has no way of doing so.