Stefan's patch:
The attached patches let x86 catch up with ppc. Both, the kernel patches
and the hypervisor patches are needed for that. Now one can create
partitions using the same oh_tools. Also x86 uses the firmware device
tree, but serializes only the vio devices data to pass them to the
kernel. Since I wanted to reuse as much code as possible I started to use the
#ifdef __i386__ and __PPC__ defines in hype_run_lpar.c .